+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 4.1.2 (TBD)
 
-* TBD
+* Shamos-Hoey simple polygon detection bug fix
 
 ## [4.1.1](https://github.com/ngageoint/simple-features-ios/releases/tag/4.1.1) (01-09-2023)